+++ src/main/java/wooteco/subway/config/WebConfig.java
@@ -6,8 +6,11 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
+
     @Override
     public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**").allowedMethods("*").allowedOriginPatterns("*");
+        registry.addMapping("/**")
+                .allowedMethods("*")
+                .allowedOriginPatterns("*");
     }
 }
