+++ src/main/java/wooteco/subway/config/auth/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package wooteco.subway.config;
+package wooteco.subway.config.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
