+++ src/main/java/wooteco/subway/config/auth/AuthenticationPrincipalArgumentResolver.java
@@ -1,19 +1,18 @@
-package wooteco.subway.config;
+package wooteco.subway.config.auth;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import wooteco.subway.dto.LoginMember;
-import wooteco.subway.infrastructure.AuthorizationExtractor;
-import wooteco.subway.infrastructure.JwtTokenProvider;
+import wooteco.subway.config.auth.infrastructure.AuthorizationExtractor;
 import wooteco.subway.service.AuthService;
 
 import javax.servlet.http.HttpServletRequest;
 
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
-    private AuthService authService;
+
+    private final AuthService authService;
 
     public AuthenticationPrincipalArgumentResolver(AuthService authService) {
         this.authService = authService;
@@ -24,13 +23,10 @@ public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArg
         return parameter.hasParameterAnnotation(AuthenticationPrincipal.class);
     }
 
-    // parameter에 @AuthenticationPrincipal이 붙어있는 경우 동작
     @Override
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         HttpServletRequest httpServletRequest = (HttpServletRequest) webRequest.getNativeRequest();
         String token = AuthorizationExtractor.extract(httpServletRequest);
-        LoginMember loginMember = authService.parseLoginMember(token);
-
-        return loginMember;
+        return authService.parseToken(token);
     }
 }
