+++ src/main/java/wooteco/subway/config/auth/AuthenticationPrincipalConfig.java
@@ -1,7 +1,8 @@
-package wooteco.subway.config;
+package wooteco.subway.config.auth;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -11,15 +12,11 @@ import java.util.List;
 
 @Configuration
 public class AuthenticationPrincipalConfig implements WebMvcConfigurer {
+
     private final AuthService authService;
 
     public AuthenticationPrincipalConfig(AuthService authService) {
         this.authService = authService;
-    }
-
-    @Override
-    public void addArgumentResolvers(List argumentResolvers) {
-        argumentResolvers.add(createAuthenticationPrincipalArgumentResolver());
     }
 
     @Bean
@@ -30,6 +27,11 @@ public class AuthenticationPrincipalConfig implements WebMvcConfigurer {
     @Bean
     public HandlerInterceptor createHandleInterceptor() {
         return new LoginInterceptor(authService);
+    }
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(createAuthenticationPrincipalArgumentResolver());
     }
 
     @Override
