+++ src/main/java/wooteco/subway/config/auth/LoginInterceptor.java
@@ -1,13 +1,14 @@
-package wooteco.subway.config;
+package wooteco.subway.config.auth;
 
 import org.springframework.web.servlet.HandlerInterceptor;
-import wooteco.subway.infrastructure.AuthorizationExtractor;
+import wooteco.subway.config.auth.infrastructure.AuthorizationExtractor;
 import wooteco.subway.service.AuthService;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 public class LoginInterceptor implements HandlerInterceptor {
+
     private final AuthService authService;
 
     public LoginInterceptor(AuthService authService) {
@@ -17,8 +18,6 @@ public class LoginInterceptor implements HandlerInterceptor {
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
         String token = AuthorizationExtractor.extract(request);
-        System.out.println("kimun log");
-        System.out.println(token);
         authService.validate(token);
         return HandlerInterceptor.super.preHandle(request, response, handler);
     }
