+++ src/main/java/wooteco/subway/config/auth/SpringConfiguration.java
@@ -1,9 +1,11 @@
-package wooteco.subway.config;
+package wooteco.subway.config.auth;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 
+@EnableWebSecurity
 @Configuration
 public class SpringConfiguration extends WebSecurityConfigurerAdapter {
