+++ src/main/java/wooteco/subway/config/auth/infrastructure/AuthorizationExtractor.java
@@ -1,12 +1,12 @@
-package wooteco.subway.infrastructure;
+package wooteco.subway.config.auth.infrastructure;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.Enumeration;
 
 public class AuthorizationExtractor {
     public static final String AUTHORIZATION = "Authorization";
-    public static String BEARER_TYPE = "Bearer";
     public static final String ACCESS_TOKEN_TYPE = AuthorizationExtractor.class.getSimpleName() + ".ACCESS_TOKEN_TYPE";
+    public static String BEARER_TYPE = "Bearer";
 
     public static String extract(HttpServletRequest request) {
         Enumeration<String> headers = request.getHeaders(AUTHORIZATION);
