+++ src/main/java/wooteco/subway/config/auth/infrastructure/JwtTokenProvider.java
@@ -1,10 +1,8 @@
-package wooteco.subway.infrastructure;
+package wooteco.subway.config.auth.infrastructure;
 
 import io.jsonwebtoken.*;
-import jdk.internal.net.http.common.Log;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import wooteco.subway.dto.LoginMember;
 
 import java.util.Date;
 import java.util.HashMap;
@@ -24,7 +22,6 @@ public class JwtTokenProvider {
         payload.put("age", age);
         Date now = new Date();
         Date validity = new Date(now.getTime() + validityInMilliseconds);
-
         return Jwts.builder()
                 .setClaims(payload)
                 .setIssuedAt(now)
@@ -33,17 +30,16 @@ public class JwtTokenProvider {
                 .compact();
     }
 
-    public LoginMember getPayload(String token) {
-        Claims payload = Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token).getBody();
-        return new LoginMember(payload.get("id", Long.class),
-                payload.get("email", String.class),
-                payload.get("age", Integer.class));
+    public Claims getPayloads(String token) {
+        return Jwts.parser()
+                .setSigningKey(secretKey)
+                .parseClaimsJws(token)
+                .getBody();
     }
 
     public boolean validateToken(String token) {
         try {
             Jws<Claims> claims = Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token);
-
             return !claims.getBody().getExpiration().before(new Date());
         } catch (JwtException | IllegalArgumentException e) {
             throw new IllegalArgumentException("유효하지 않은 토큰입니다.");
